Add output buckets + remove Generator (#23)
bench: 8711412
