+++ src/AwesomeOpossum/Logic/Evaluation/ValueNetwork.cs
@@ -27,7 +27,7 @@ namespace AwesomeOpossum.Logic.Evaluation
         public const int INPUT_BUCKETS = 2;
         public const int INPUT_SIZE = 768;
         public const int L1_SIZE = 256;
-        public const int OUTPUT_BUCKETS = 1;
+        public const int OUTPUT_BUCKETS = 8;
 
         private const int BUCKET_DIV = ((32 + OUTPUT_BUCKETS - 1) / OUTPUT_BUCKETS);
         private const int QA = 256;
@@ -152,6 +152,48 @@ namespace AwesomeOpossum.Logic.Evaluation
             int output = SIMDBindings.ValueEvaluateFn(stmData, ntmData, l1Weights, l1Bias);
 
             return int.Clamp(output, ScoreTTLoss + 1, ScoreTTWin - 1);
+        }
+
+
+        [UnmanagedCallersOnly]
+        public static int EvaluateImpl(short* stmData, short* ntmData, short* l1Weights, short l1Bias)
+        {
+            Vector256<short> maxVec = Vector256.Create((short)QA);
+            Vector256<short> zeroVec = Vector256<short>.Zero;
+            Vector256<int> sum = Vector256<int>.Zero;
+
+            int SimdChunks = L1_SIZE / Vector256<short>.Count;
+
+            var ourData = (Vector256<short>*)stmData;
+            var theirData = (Vector256<short>*)ntmData;
+            var ourWeights = (Vector256<short>*)(&l1Weights[0]);
+            var theirWeights = (Vector256<short>*)(&l1Weights[L1_SIZE]);
+            for (int i = 0; i < SimdChunks; i++)
+            {
+                Vector256<short> clamp = Vector256.Min(maxVec, Vector256.Max(zeroVec, ourData[i]));
+                Vector256<short> mult = clamp * ourWeights[i];
+
+                (var mLo, var mHi) = Vector256.Widen(mult);
+                (var cLo, var cHi) = Vector256.Widen(clamp);
+
+                sum = Vector256.Add(sum, Vector256.Add(mLo * cLo, mHi * cHi));
+            }
+
+            for (int i = 0; i < SimdChunks; i++)
+            {
+                Vector256<short> clamp = Vector256.Min(maxVec, Vector256.Max(zeroVec, theirData[i]));
+                Vector256<short> mult = clamp * theirWeights[i];
+
+                (var mLo, var mHi) = Vector256.Widen(mult);
+                (var cLo, var cHi) = Vector256.Widen(clamp);
+
+                sum = Vector256.Add(sum, Vector256.Add(mLo * cLo, mHi * cHi));
+            }
+
+            int output = Vector256.Sum(sum);
+            output = (output / QA) + l1Bias;
+
+            return output * OUTPUT_SCALE / (QA * QB);
         }
