--- src/Generators/Generators.csproj
@@ -1,26 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <EnforceExtendedAnalyzerRules>true</EnforceExtendedAnalyzerRules>
-    
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <GenerateDependencyFile>false</GenerateDependencyFile>
-    <GenerateTargetFrameworkAttribute>false</GenerateTargetFrameworkAttribute>
-    
-    <InvariantGlobalization>true</InvariantGlobalization>
-    <IncludeSatelliteAssembliesForPublishedRuntime>false</IncludeSatelliteAssembliesForPublishedRuntime>
-    <SatelliteResourceLanguages>en</SatelliteResourceLanguages>
-    <Platforms>x64</Platforms>
-
-    <EmitCompilerGeneratedFiles>true</EmitCompilerGeneratedFiles>
-    <CompilerGeneratedFilesOutputPath>$(BaseIntermediateOutputPath)tmp</CompilerGeneratedFilesOutputPath>
-    
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.CodeAnalysis.Common" Version="4.13.0" />
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.13.0" PrivateAssets="all" />
-  </ItemGroup>
-
-</Project>
