+++ value.txt
@@ -1,1 +1,1 @@
-opv-008-gini
+opv-008
